MEGAFEAT: NOW WITH HIGH SECURITY

+++ app/routes/note_routes.js
@@ -1,4 +1,6 @@
 const fetch = require("node-fetch");
+let realToken = '';
+let decoyToken = 'Acc̲̥̖̲̹̀͛͛̚̚es̺̉s̀͟_T͍͖̀͞o͈̘̦̽͌̃k̰͠e̢͗͆͟n̢̨̼̺̫̉̓̆͂̄_̤̬͚̇̄͘?̠̏_̟̌D̪͡R̮̲̖̺͙͛̋̄̃̕EA̝̦̤̹͊͑͆͘M̦͘_̪̲͒̆Í̡͎͚͙̟̍̈́̒͡T͗ͅ_XD'
 
 
 module.exports = function (app, db) {
@@ -13,18 +15,23 @@ module.exports = function (app, db) {
                 })
             })
                 .then(response => response.json())
-                .then(data => res.send(JSON.stringify(data)))
+                .then(data => {
+                    realToken = data.access_token;
+                    data.access_token = decoyToken
+                    res.send(JSON.stringify(data))
+                })
         } catch (e) {
             res.send(e)
         }
     })
     app.post('/api/product', async (req, res) => {
         try {
+            if( ! req.body.access_token === decoyToken ) res.send( 'you need to use my token U.U \n' + decoyToken)
             await fetch('https://striveschool-api.herokuapp.com/api/product/', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${req.body['access_token']}`,
+                    'Authorization': `Bearer ${realToken}`,
                 },
                 body: JSON.stringify(req.body)
             })
@@ -36,12 +43,13 @@ module.exports = function (app, db) {
         }
     })
     app.post('/api/product/all', async (req, res) => {
+        if( ! req.body.access_token === decoyToken ) res.send( 'you need to use my token U.U \n' + decoyToken)
         try {
             await fetch(`https://striveschool-api.herokuapp.com/api/product`, {
                 method: 'GET',
                 headers: {
                     'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${req.body['access_token']}`,
+                    'Authorization': `Bearer ${realToken}`,
                 },
             })
                 .then(response => response.json())
@@ -51,52 +59,55 @@ module.exports = function (app, db) {
             res.send(e)
         }
     })
-    /*
-    const {createProxyMiddleware} = require('http-proxy-middleware')
-    app.use('/api/imageUpload', createProxyMiddleware({
-        target: `https://api.imgbb.com/1/upload?key=${process.env.IMGBB_TOKEN}`,
-        changeOrigin: true
-    }), () =>{ console.log( )})*/
+
     let imgbbUploader = require("imgbb-uploader");
 
     app.post('/api/imageUpload', async (req, res) => {
+        if( ! req.body.access_token === decoyToken ) res.send( 'you need to use my token U.U \n' + decoyToken)
             try {
                 imgbbUploader({
                     apiKey: process.env.IMGBB_TOKEN,
                     base64string: req.body['blob'].split(',')[1]
                 })
-                    .then((response) => res.send( response) )
+                    .then((response) => res.send(response))
                     .catch((error) => error)
-               /* return await fetch(`https://api.imgbb.com/1/upload?key=${process.env.IMGBB_TOKEN}`, {
-                    method: 'POST',
-                    body: req.body,
-                })
-                    .then(response => {
-                        console.log(response);
-                        response.json()
-                    })
-                    .then(data => {
-                        console.log(data);
-                        res.send(data)
-                    }) */
+                /* return await fetch(`https://api.imgbb.com/1/upload?key=${process.env.IMGBB_TOKEN}`, {
+                     method: 'POST',
+                     body: req.body,
+                 })
+                     .then(response => {
+                         console.log(response);
+                         response.json()
+                     })
+                     .then(data => {
+                         console.log(data);
+                         res.send(data)
+                     }) */
             } catch (e) {
                 console.log(e)
                 res.send(e)
             }
         }
     )
-    app.delete('/api/product/:id',async (req, res) =>{
-        console.log(req.body)
+    app.delete('/api/product/:id', async (req, res) => {
+        if( ! req.body.access_token === decoyToken ) res.send( 'you need to use my token U.U \n' + decoyToken)
         try {
-            await fetch(`https://striveschool-api.herokuapp.com/api/product/${req.params.id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${req.body['access_token']}`,
-                },
-            })
-                .then(response => response.json())
-                .then(data => { console.log(data); res.send(data)  })// res.send(data))
+            if (req.params.id === '60b04910dc14580015e4ad61') {
+                res.send('Nice try dude <3 the is the first think of everyone ')
+            } else {
+                await fetch(`https://striveschool-api.herokuapp.com/api/product/${req.params.id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Authorization': `Bearer ${realToken}`,
+                    },
+                })
+                    .then(response => response.json())
+                    .then(data => {
+                        console.log(data);
+                        res.send(data)
+                    })
+            }
 
         } catch (e) {
             res.send(e)
